% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_teams.R
\docType{data}
\name{distinct_teams}
\alias{distinct_teams}
\title{Distinct Team Names}
\format{
A character vector with 560 elements:
}
\usage{
distinct_teams
}
\description{
A dataset containing distinct team names for women's lacorsse teams
in the NCAAA
}
\keyword{datasets}
