% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncaa_wlax_team_roster.R
\name{get_ncaa_wlax_team_roster}
\alias{get_ncaa_wlax_team_roster}
\title{Get Team Rosters for Women's Lacrosse}
\usage{
get_ncaa_wlax_team_roster(team_id, year)
}
\arguments{
\item{team_id}{The unique team id from the NCAA}

\item{year}{The season for which you want rsoters
If you want the 2017-18 season, you would use 2018.}
}
\value{
A data frame
}
\description{
Get Team Rosters for Women's Lacrosse
}
\examples{
\dontrun{get_ncaa_wlax_team_roster(392, 2019)}
}
