% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_wlax_ncaa_team_lu.R
\docType{data}
\name{master_wlax_ncaa_team_lu}
\alias{master_wlax_ncaa_team_lu}
\title{Master NCAA Team Lookup}
\format{
A data frame
}
\usage{
master_wlax_ncaa_team_lu
}
\description{
A data frame containing a master list of women's lacrosse teams, team IDs, conferences, and
conference IDs
}
\keyword{datasets}
