% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncaa_stats_year_lu_table.R
\docType{data}
\name{ncaa_stats_year_lu_table}
\alias{ncaa_stats_year_lu_table}
\title{NCAA Women's Stats and Season Lookup Table}
\format{
A data frame
}
\usage{
ncaa_stats_year_lu_table
}
\description{
A data frame containing a master list of women's lacrosse stats and season IDs
}
\keyword{datasets}
